Move cloudinit module into internal
CABPK is not willing to support cloudinit as an importable library.
