+++ bootstrap/kubeadm/internal/cloudinit/cloudinit.go
