+++ bootstrap/kubeadm/internal/cloudinit/commands.go
