+++ bootstrap/kubeadm/internal/cloudinit/controlplane_join.go
