+++ bootstrap/kubeadm/internal/cloudinit/node.go
