+++ bootstrap/kubeadm/internal/cloudinit/utils.go
