+++ docs/book/src/providers/v1alpha2-to-v1alpha3.md
@@ -105,7 +105,6 @@
 - The slice in Cluster.Status has been removed and replaced by a single APIEndpoint field under Spec.
 - Infrastructure providers MUST expose a ControlPlaneEndpoint field in their cluster infrastructure resource at `Spec.ControlPlaneEndpoint`. They may optionally remove the `Status.APIEndpoints` field (Cluster API no longer uses it).
 
-
 ## Data generated from a bootstrap provider is now stored in a secret.
 
 - The Cluster API Machine Controller no longer reconciles the bootstrap provider `status.bootstrapData` field, but instead looks at `status.dataSecretName`.
@@ -113,3 +112,8 @@
 - Bootstrap providers must create a Secret in the bootstrap resource's namespace and store the name in the bootstrap resource's `status.dataSecretName` field.
     - On reconciliation, we suggest to migrate from the deprecated field to a secret reference.
 - Infrastructure providers must look for the bootstrap data secret name in `Machine.Spec.Bootstrap.DataSecretName` and fallback to `Machine.Spec.Bootstrap.Data`.
+
+## The `cloudinit` module under the Kubeadm bootstrap provider has been made private
+
+The `cloudinit` module has been moved to an `internal` directory as it is not designed to be a public interface consumed
+outside of the existing module.
